requirements.txt
# Core dependencies
playwright==1.49.1
asyncio==3.4.3
groq==0.13.0

# Data handling
urllib3==2.3.0

# Development dependencies
python-dotenv==1.0.1

# FastAPI dependencies for visualizer
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12